feat(lists): Show playable icon in thumbnail for list items
