+++ shared/src/api/queries/entityLists/gql/ListItemFragment.graphql
@@ -7,6 +7,7 @@ fragment ListItemFragment on BaseNode {
     taskType
     assignees
     ownAttrib
+    hasReviewables
     folder {
       path
       folderType
@@ -20,6 +21,7 @@ fragment ListItemFragment on BaseNode {
     folderType
     path
     ownAttrib
+    hasReviewables
   }
   ... on ProductNode {
     name
@@ -35,6 +37,7 @@ fragment ListItemFragment on BaseNode {
     name
     status
     tags
+    hasReviewables
     product {
       id
       name
