+++ src/pages/ProjectListsPage/context/ListItemsDataContext.tsx
@@ -16,6 +16,7 @@ import useDeleteListItems, { UseDeleteListItemsReturn } from '../hooks/useDelete
 import { ContextMenuItemConstructors } from '@shared/containers/ProjectTreeTable/hooks/useCellContextMenu'
 import { useEntityListsContext } from './EntityListsContext'
 import useReorderListItem, { UseReorderListItemReturn } from '../hooks/useReorderListItem'
+import useBuildListItemsTableData from '../hooks/useBuildListItemsTableData'
 
 export type ListItemsMap = Map<string, EntityListItem>
 
@@ -129,53 +130,6 @@ export const ListItemsDataProvider = ({ children }: ListItemsDataProviderProps) 
     return new Map(listItemsData.map((item) => [item.id, item]))
   }, [listItemsData])
 
-  const extractPath = (item: EntityListItem, entityType: string): string => {
-    switch (entityType) {
-      case 'folder':
-        return item.path || ''
-      case 'task':
-        return item.folder?.path || ''
-      case 'product':
-        return item.folder?.path || ''
-      case 'version':
-        return item.product?.folder?.path || '' + item.task?.name || ''
-      default:
-        return ''
-    }
-  }
-
-  const extractSubTypes = (
-    item: EntityListItem,
-    entityType?: string,
-  ): {
-    subType?: string
-    folderType?: string
-    taskType?: string
-    productType?: string
-  } => {
-    switch (entityType) {
-      case 'folder':
-        return { subType: item.folderType, folderType: item.folderType }
-      case 'task':
-        return {
-          subType: item.taskType,
-          taskType: item.taskType,
-          folderType: item.folder?.folderType,
-        }
-      case 'product':
-        return { subType: item.productType || '', folderType: item.folder?.folderType }
-      case 'version':
-        return {
-          subType: undefined,
-          productType: item.product?.productType,
-          folderType: item.product?.folder?.folderType,
-          taskType: item.task?.taskType,
-        }
-      default:
-        return {}
-    }
-  }
-
   // filter out attribFields by scope
   const scopedAttribFields = useMemo(
     () =>
@@ -185,33 +139,12 @@ export const ListItemsDataProvider = ({ children }: ListItemsDataProviderProps) 
     [attribFields, selectedList?.entityType],
   )
 
-  // convert listItemsData into tableData
-  const listItemsTableData = useMemo(() => {
-    const tableRows: TableRow[] = listItemsData.map((item) => ({
-      id: item.id,
-      name: item.name,
-      label:
-        (item.entityType === 'version' ? `${item.product?.name} - ` : '') +
-        (item.label || item.name),
-      entityId: item.entityId,
-      entityType: item.entityType,
-      assignees: item.assignees || [],
-      ...extractSubTypes(item, item.entityType), // subType, folderType, taskType, productType
-      updatedAt: item.updatedAt,
-      attrib: item.attrib,
-      ownAttrib: item.ownAttrib
-        ? [...item.ownAttrib, ...item.ownItemAttrib]
-        : Object.keys(item.attrib), // not all types use ownAttrib so fallback to attrib keys
-      icon: getEntityTypeData(item.entityType, extractSubTypes(item, item.entityType).subType)
-        ?.icon,
-      path: extractPath(item, item.entityType),
-      tags: item.tags,
-      status: item.status,
-      subRows: [],
-    }))
+  console.log(listItemsData)
 
-    return tableRows
-  }, [listItemsData])
+  // convert listItemsData into tableData
+  const listItemsTableData = useBuildListItemsTableData({
+    listItemsData,
+  })
 
   const foldersMap: FolderNodeMap = new Map(
     // @ts-ignore
